+++ src/collective/big/bang/tests/test_setup.py
@@ -1,7 +1,7 @@
 # -*- coding: utf-8 -*-
 """Setup tests for this package."""
-from collective.bigbang.testing import (
-    COLLECTIVE_BIGBANG_INTEGRATION_TESTING  # noqa: E501,
+from collective.big.bang.testing import (
+    COLLECTIVE_BIG_BANG_INTEGRATION_TESTING  # noqa: E501,
 )
 from plone import api
 from plone.app.testing import setRoles, TEST_USER_ID
@@ -16,9 +16,9 @@ except ImportError:
 
 
 class TestSetup(unittest.TestCase):
-    """Test that collective.bigbang is properly installed."""
+    """Test that collective.big.bang is properly installed."""
 
-    layer = COLLECTIVE_BIGBANG_INTEGRATION_TESTING
+    layer = COLLECTIVE_BIG_BANG_INTEGRATION_TESTING
 
     def setUp(self):
         """Custom shared utility setup for tests."""
@@ -29,13 +29,13 @@ class TestSetup(unittest.TestCase):
             self.installer = api.portal.get_tool('portal_quickinstaller')
 
     def test_product_installed(self):
-        """Test if collective.bigbang is installed."""
+        """Test if collective.big.bang is installed."""
         self.assertTrue(self.installer.isProductInstalled(
-            'collective.bigbang'))
+            'collective.big.bang'))
 
     def test_browserlayer(self):
         """Test that ICollectiveBigbangLayer is registered."""
-        from collective.bigbang.interfaces import (
+        from collective.big.bang.interfaces import (
             ICollectiveBigbangLayer)
         from plone.browserlayer import utils
         self.assertIn(
@@ -45,7 +45,7 @@ class TestSetup(unittest.TestCase):
 
 class TestUninstall(unittest.TestCase):
 
-    layer = COLLECTIVE_BIGBANG_INTEGRATION_TESTING
+    layer = COLLECTIVE_BIG_BANG_INTEGRATION_TESTING
 
     def setUp(self):
         self.portal = self.layer['portal']
@@ -55,17 +55,17 @@ class TestUninstall(unittest.TestCase):
             self.installer = api.portal.get_tool('portal_quickinstaller')
         roles_before = api.user.get_roles(TEST_USER_ID)
         setRoles(self.portal, TEST_USER_ID, ['Manager'])
-        self.installer.uninstallProducts(['collective.bigbang'])
+        self.installer.uninstallProducts(['collective.big.bang'])
         setRoles(self.portal, TEST_USER_ID, roles_before)
 
     def test_product_uninstalled(self):
-        """Test if collective.bigbang is cleanly uninstalled."""
+        """Test if collective.big.bang is cleanly uninstalled."""
         self.assertFalse(self.installer.isProductInstalled(
-            'collective.bigbang'))
+            'collective.big.bang'))
 
     def test_browserlayer_removed(self):
         """Test that ICollectiveBigbangLayer is removed."""
-        from collective.bigbang.interfaces import \
+        from collective.big.bang.interfaces import \
             ICollectiveBigbangLayer
         from plone.browserlayer import utils
         self.assertNotIn(